refactor: extract now-playing pure builders and modals into modules
Move buildEditNoteModal, buildEditNotesModal, and buildNowPlayingAddModal
into now-playing/nowPlayingModals.ts, and the pure component builders
(remove, edit-platform, sort) into nowPlayingListRenderer.ts. These are
zero-state pure functions; call sites updated from this.X to imports.

Refs #894
